schema: separated 2525-C/SKKM.
